Update correction
Had to make correction to count of image nodes to prevent the
application from throwing an exception if the user tries to go past the
available number of pictures. Also, the count of the images in the NASA
provided iotdxml.xml doc, is off by 2, which is how I found the
exception.