Internet exception
Added check in the retrieval method to check if there is an internet
connection before attempting to download. Previously there were checks
put in to prevent the application from launching if internet access was
unavailable. But, if the application is running and internet access is
lost, a .Net exception is triggered